GIF to video: let the Image block own the conversion helpers
Per review feedback, a single block should handle the transformation.
Move the helpers from the shared utils folder into the Image block's
transforms module, and drop their unit tests: exercising the Block API
is not the block's responsibility, and the end-to-end test covers the
conversion behavior.
