+++ packages/block-library/src/image/test/transforms.js
@@ -1,29 +1,7 @@
 /**
- * WordPress dependencies
- */
-import { select } from '@wordpress/data';
-
-/**
  * Internal dependencies
  */
-import transforms, { stripFirstImage } from '../transforms';
-
-jest.mock( '@wordpress/data', () => ( {
-	select: jest.fn(),
-} ) );
-
-jest.mock( '@wordpress/core-data', () => ( {
-	store: 'core',
-} ) );
-
-/*
- * core/video is not registered in this unit test, so stub createBlock with a
- * lightweight factory that preserves the name and attributes for assertion.
- */
-jest.mock( '@wordpress/blocks', () => ( {
-	createBlock: jest.fn( ( name, attributes ) => ( { name, attributes } ) ),
-	getBlockAttributes: jest.fn(),
-} ) );
+import { stripFirstImage } from '../transforms';
 
 describe( 'stripFirstImage', () => {
 	test( 'should do nothing if no image is present', () => {
@@ -84,78 +62,5 @@ describe( 'stripFirstImage', () => {
 				{ shortcode: { content: '<p><a><img></a></p><p><img></p>' } }
 			)
 		).toEqual( '<p><img></p>' );
-	} );
-} );
-
-describe( 'animated GIF to Video block transform', () => {
-	const toVideo = transforms.to.find( ( t ) =>
-		t.blocks.includes( 'core/video' )
-	);
-	const GIF_URL = 'https://example.com/wp-content/uploads/cat.gif';
-	let getEntityRecord;
-
-	const companionRecord = {
-		source_url: GIF_URL,
-		media_details: {
-			animated_video: 'cat.mp4',
-			animated_video_poster: 'cat.jpg',
-			width: 320,
-			height: 240,
-		},
-	};
-
-	beforeEach( () => {
-		jest.clearAllMocks();
-		getEntityRecord = jest.fn( () => companionRecord );
-		select.mockReturnValue( { getEntityRecord } );
-	} );
-
-	test( 'matches only when the attachment has a video companion', () => {
-		expect( toVideo.isMatch( { id: 7, url: GIF_URL } ) ).toBe( true );
-
-		// Not a GIF.
-		expect(
-			toVideo.isMatch( {
-				id: 7,
-				url: 'https://example.com/wp-content/uploads/cat.jpg',
-			} )
-		).toBe( false );
-
-		// No attachment id (e.g. an external image).
-		expect( toVideo.isMatch( { url: GIF_URL } ) ).toBe( false );
-
-		// A GIF without a sideloaded companion video.
-		getEntityRecord.mockReturnValue( {
-			source_url: GIF_URL,
-			media_details: { width: 320, height: 240 },
-		} );
-		expect( toVideo.isMatch( { id: 7, url: GIF_URL } ) ).toBe( false );
-	} );
-
-	test( 'creates a GIF-behaving video block from the companion', () => {
-		const block = toVideo.transform( {
-			id: 7,
-			url: GIF_URL,
-			caption: 'A cat',
-			align: 'wide',
-			anchor: 'my-gif',
-		} );
-
-		expect( block.name ).toBe( 'core/video' );
-		expect( block.attributes ).toEqual( {
-			id: 7,
-			src: 'https://example.com/wp-content/uploads/cat.mp4',
-			poster: 'https://example.com/wp-content/uploads/cat.jpg',
-			caption: 'A cat',
-			controls: false,
-			loop: true,
-			autoplay: true,
-			muted: true,
-			playsInline: true,
-			width: 320,
-			height: 240,
-			align: 'wide',
-			anchor: 'my-gif',
-		} );
 	} );
 } );
